meme_tracker/requirements.txt
python-telegram-bot==20.0  # 使用v20稳定版本
tweepy==4.14.0
praw==7.7.1
aiohttp==3.9.5
ratelimit==2.2.1
python-dotenv==1.0.1
